Ahora cada Conductor podrá ver el viaje asignado, y marcar el si el viaje ha terminado

+++ GestionDeBuses/database/migrations/2024_12_04_190221_add_user_id_to_buses_table.php
@@ -12,8 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('buses', function (Blueprint $table) {
-            //
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            // Verifica si la columna 'user_id' no existe antes de agregarla
+            if (!Schema::hasColumn('buses', 'user_id')) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            }
         });
     }
 
@@ -23,6 +25,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('buses', function (Blueprint $table) {
+            // Elimina la clave foránea y la columna
             $table->dropForeign(['user_id']);
             $table->dropColumn('user_id');
         });
